requirements.txt
django==3.1.2
djangorestframework==3.12.1
environs==8.0.0
uritemplate==3.0.1